+++ features/merchant/shop_score/src/main/java/com/tokopedia/shop/score/view/model/ShopScoreDetailItem.kt
@@ -2,8 +2,8 @@ package com.tokopedia.shop.score.view.model
 
 data class ShopScoreDetailItem(
     var title: String?,
-    val value: Int,
-    val maxValue: Int,
+    val value: Float,
+    val maxValue: Float,
     val description: String?,
     val progressBarColor: String?
 )
